+++ styles/PostStyle.tsx
